+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Editor: Show cursor all the time.
 * Breakpoints: Allow multiple filename:line entries for the --break-source option.
 * Control: Add ^D to toggle record playback direction for RR, Undo, and gdb's Record.
+* Control: Add SHIFT+F5 (etal) to reverse step (etc.) during record playback for RR, Undo, and gdb's Record.
 
 ## [2.6] - 2025-10-07
 * Watchpoints: Fixed regression when adding commands to a watchpoint.